Fix /test_llama_manager.py
