+++ services/chat_service/tests/test_llama_manager.py
@@ -1,4 +1,3 @@
-from unittest import mock
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
@@ -40,209 +39,6 @@ def manager(mock_tools, mock_subagents):
     )
 
 
-@pytest.mark.asyncio
-@patch("services.chat_service.llama_manager.history_manager")
-@patch("services.chat_service.llama_manager.context_module")
-async def test_get_memory_selects_relevant_messages(
-    mock_context, mock_history, manager
-):
-    # Setup
-    fake_messages = [
-        MagicMock(
-            model_dump=MagicMock(return_value={"user_id": "user42", "content": "hi"})
-        ),
-        MagicMock(
-            model_dump=MagicMock(
-                return_value={"user_id": "assistant", "content": "hello"}
-            )
-        ),
-    ]
-    mock_history.get_thread_history = AsyncMock(return_value=fake_messages)
-
-    # Mock the async function
-    async def mock_dynamic_selection(*args, **kwargs):
-        return [
-            {"user_id": "user42", "content": "hi"},
-            {"user_id": "assistant", "content": "hello"},
-        ]
-
-    mock_context.dynamic_context_selection = mock_dynamic_selection
-
-    # Test
-    result = await manager.get_memory("test query")
-
-    # Verify
-    assert len(result) == 2
-    assert result[0]["content"] == "hi"
-    assert result[0]["user_id"] == "user42"
-    assert result[1]["content"] == "hello"
-    assert result[1]["user_id"] == "assistant"
-    mock_history.get_thread_history.assert_awaited_once_with(123, limit=100)
-
-
-@pytest.mark.asyncio
-@patch("services.chat_service.llama_manager.FunctionTool")
-@patch("services.chat_service.llama_manager.Memory")
-@patch("services.chat_service.llama_manager.FunctionCallingAgent")
-@patch("services.chat_service.llama_manager.history_manager")
-@patch("services.chat_service.llama_manager.context_module")
-async def test_build_agent_constructs_agent(
-    mock_context,
-    mock_history,
-    mock_agent_cls,
-    mock_memory,
-    mock_tool,
-    manager,
-    mock_tools,
-    mock_subagents,
-):
-    # Setup
-    fake_messages = [
-        {"user_id": "user42", "content": "hi"},
-        {"user_id": "assistant", "content": "hello"},
-    ]
-
-    # Mock history manager
-    mock_history.get_thread_history = AsyncMock(return_value=[])
-
-    # Mock context module
-    async def mock_dynamic_selection(*args, **kwargs):
-        return fake_messages
-
-    mock_context.dynamic_context_selection = mock_dynamic_selection
-
-    # Mock tools and memory
-    mock_tool.from_defaults.side_effect = lambda fn: f"tool-{fn.__name__}"
-    mock_memory.from_defaults.return_value = MagicMock()
-
-    # Mock agent
-    mock_agent_instance = AsyncMock()
-    mock_agent_cls.from_tools.return_value = mock_agent_instance
-
-    # Run
-    await manager.build_agent("hi")
-
-    # Assert FunctionCallingAgent was used with tools
-    mock_agent_cls.from_tools.assert_called_once()
-    assert manager.agent == mock_agent_instance
-    # Should wrap all tools and subagents
-    assert len(mock_tool.from_defaults.mock_calls) == len(mock_tools) + len(
-        mock_subagents
-    )
-    mock_agent_cls.from_tools.assert_called_once()
-    mock_memory.from_defaults.assert_called_once()
-    args, kwargs = mock_memory.from_defaults.call_args
-    chat_history = kwargs.get("chat_history") or args[0]
-    # Check user and assistant messages
-    assert chat_history[0].role == "user"
-    assert chat_history[0].content == "hi"
-    assert chat_history[1].role == "assistant"
-    assert chat_history[1].content == "hello"
-
-
-@pytest.mark.asyncio
-@patch("services.chat_service.llama_manager.FunctionTool")
-@patch("services.chat_service.llama_manager.Memory")
-@patch("services.chat_service.llama_manager.FunctionCallingAgent")
-@patch("services.chat_service.llama_manager.history_manager")
-@patch("services.chat_service.llama_manager.context_module")
-@patch("services.chat_service.llama_manager.llm_manager.get_llm")
-async def test_chat_calls_agent_and_appends_history(
-    mock_get_llm,
-    mock_context,
-    mock_history,
-    mock_agent,
-    mock_memory,
-    mock_tool,
-    manager,
-):
-    # Setup mock thread
-    mock_thread = MagicMock()
-    mock_thread.id = 123
-    mock_history.Thread.objects.get = AsyncMock(return_value=mock_thread)
-
-    # Patch async methods with AsyncMock
-    mock_history.append_message = AsyncMock()
-    mock_history.get_thread_history = AsyncMock(return_value=[])
-    mock_history.create_thread = AsyncMock(return_value=mock_thread)
-
-    # Mock the LLM
-    mock_llm = AsyncMock()
-    mock_llm.achat.return_value = "[FAKE LLM RESPONSE] You said: hello world"
-    mock_get_llm.return_value = mock_llm
-
-    # Setup agent
-    fake_agent = MagicMock()
-    fake_agent.achat = AsyncMock(
-        return_value="[FAKE LLM RESPONSE] You said: hello world"
-    )
-    manager.agent = fake_agent
-    manager.user_id = "user42"
-    manager.thread_id = 123
-
-    # Run
-    result = await manager.chat("hello world")
-
-    # Assert
-    assert result == "[FAKE LLM RESPONSE] You said: hello world"
-    # In the current implementation, user message is appended after building the agent
-    # So we just check that the assistant's response was appended
-    mock_history.append_message.assert_called_once_with(
-        123, "assistant", "[FAKE LLM RESPONSE] You said: hello world"
-    )
-
-
-@pytest.mark.asyncio
-@patch("services.chat_service.llama_manager.history_manager")
-@patch("services.chat_service.llama_manager.context_module")
-@patch("services.chat_service.llama_manager.llm_manager.get_llm")
-async def test_chat_builds_agent_if_none(
-    mock_get_llm, mock_context, mock_history, manager, mock_tools
-):
-    # Setup - no tools or subagents for this test
-    manager.agent = None
-    manager.tools = []
-    manager.subagents = []
-
-    # Create a proper FakeLLM instance
-    from services.chat_service.llm_manager import FakeLLM
-
-    fake_llm = FakeLLM()
-
-    # Mock the LLM manager to return our fake LLM
-    mock_get_llm.return_value = fake_llm
-
-    # Mock thread
-    mock_thread = MagicMock()
-    mock_thread.id = 123
-
-    # Create a mock Thread model class
-    mock_thread_model = MagicMock()
-    mock_thread_model.objects.get = AsyncMock(return_value=mock_thread)
-
-    # Mock history manager
-    mock_history.Thread = mock_thread_model
-    mock_history.create_thread = AsyncMock(return_value=mock_thread)
-    mock_history.append_message = AsyncMock()
-    mock_history.get_thread_history = AsyncMock(return_value=[])
-
-    # Mock context module
-    async def mock_dynamic_selection(*args, **kwargs):
-        return []
-
-    mock_context.dynamic_context_selection = mock_dynamic_selection
-
-    # Test
-    result = await manager.chat("test input")
-
-    # Verify the response is as expected from FakeLLM
-    assert "test input" in result
-
-    # In FakeLLM case, we don't actually build an agent
-    # Just verify the message was appended to history
-    mock_history.append_message.assert_any_call(123, "assistant", mock.ANY)
-
-
 def test_init_defaults():
     manager = ChatAgentManager(thread_id=1, user_id="u")
     assert manager.thread_id == 1
@@ -252,3 +48,217 @@ def test_init_defaults():
     assert manager.subagents == []
     assert manager.agent is None
     assert manager.memory is None
+
+
+@pytest.fixture
+def orchestration_tools():
+    """Sample tools for orchestration testing."""
+
+    def calendar_tool():
+        """Get calendar information."""
+        return "Calendar: Meeting at 3pm"
+
+    def email_tool():
+        """Send emails."""
+        return "Email sent successfully"
+
+    def weather_tool():
+        """Get weather information."""
+        return "Weather: Sunny, 72°F"
+
+    return [calendar_tool, email_tool, weather_tool]
+
+
+@pytest.fixture
+def sample_subagents():
+    """Sample subagent configurations."""
+    return [
+        {"name": "email_specialist", "tools": ["email_tool"]},
+        {"name": "calendar_specialist", "tools": ["calendar_tool"]},
+    ]
+
+
+@pytest.mark.asyncio
+async def test_orchestration_initialization(orchestration_tools, sample_subagents):
+    """Test that the orchestration layer initializes correctly."""
+    manager = ChatAgentManager(
+        thread_id=100,
+        user_id="orchestration_user",
+        tools=orchestration_tools,
+        subagents=sample_subagents,
+        max_tokens=4096,
+    )
+
+    # Check basic properties
+    assert manager.thread_id == 100
+    assert manager.user_id == "orchestration_user"
+    assert manager.max_tokens == 4096
+
+    # Check tool distribution
+    assert len(manager.tools) == 3
+    assert len(manager.subagents) == 2
+
+    # Check tool distribution strategy
+    assert "main_agent" in manager.tool_distribution
+    assert "specialized_agents" in manager.tool_distribution
+    assert (
+        len(manager.tool_distribution["main_agent"]) == 3
+    )  # All tools go to main for now
+
+    # Check initial state
+    assert len(manager.active_agents) == 0
+    assert manager.main_agent is None
+
+
+@pytest.mark.asyncio
+@patch("services.chat_service.chat_agent.history_manager")
+async def test_main_agent_creation(mock_history, orchestration_tools):
+    """Test that the main agent is created correctly."""
+    mock_history.get_thread_history.return_value = []
+
+    manager = ChatAgentManager(
+        thread_id=101,
+        user_id="main_agent_user",
+        tools=orchestration_tools,
+    )
+
+    # Create main agent
+    main_agent = await manager._create_main_agent()
+
+    # Verify agent was created and registered
+    assert main_agent is not None
+    assert manager.main_agent is main_agent
+    assert "main" in manager.active_agents
+    assert manager.active_agents["main"] is main_agent
+
+    # Verify agent has the tools
+    assert len(main_agent.tools) == 3
+
+
+@pytest.mark.asyncio
+@patch("services.chat_service.chat_agent.history_manager")
+async def test_subagent_creation(mock_history, orchestration_tools, sample_subagents):
+    """Test that sub-agents are created correctly."""
+    mock_history.get_thread_history.return_value = []
+
+    manager = ChatAgentManager(
+        thread_id=102,
+        user_id="subagent_user",
+        tools=orchestration_tools,
+        subagents=sample_subagents,
+    )
+
+    # Create sub-agents
+    subagents = await manager._create_subagents()
+
+    # Verify sub-agents were created
+    assert len(subagents) == 2
+    assert "subagent_0" in subagents
+    assert "subagent_1" in subagents
+
+    # Verify they're registered in active_agents
+    assert "subagent_0" in manager.active_agents
+    assert "subagent_1" in manager.active_agents
+
+    # Verify sub-agents have different memory configuration
+    subagent_0 = subagents["subagent_0"]
+    assert not subagent_0.enable_fact_extraction  # Sub-agents don't extract facts
+    assert subagent_0.enable_vector_memory  # But they can search history
+
+
+@pytest.mark.asyncio
+@patch("services.chat_service.chat_agent.history_manager")
+async def test_orchestration_build_agent(
+    mock_history, orchestration_tools, sample_subagents
+):
+    """Test building the full orchestrated system."""
+    mock_history.get_thread_history.return_value = []
+
+    manager = ChatAgentManager(
+        thread_id=103,
+        user_id="build_user",
+        tools=orchestration_tools,
+        subagents=sample_subagents,
+    )
+
+    # Build the system
+    await manager.build_agent()
+
+    # Verify main agent was created
+    assert manager.main_agent is not None
+    assert "main" in manager.active_agents
+
+    # Verify sub-agents were created
+    assert len(manager.active_agents) == 3  # main + 2 subagents
+    assert "subagent_0" in manager.active_agents
+    assert "subagent_1" in manager.active_agents
+
+
+@pytest.mark.asyncio
+async def test_query_routing():
+    """Test query routing logic."""
+    manager = ChatAgentManager(
+        thread_id=104,
+        user_id="routing_user",
+    )
+
+    # Test routing (currently always returns "main")
+    target = await manager._route_query("What's the weather?")
+    assert target == "main"
+
+    target = await manager._route_query("Send an email")
+    assert target == "main"
+
+
+@pytest.mark.asyncio
+async def test_memory_aggregation():
+    """Test memory information aggregation from multiple agents."""
+    manager = ChatAgentManager(
+        thread_id=105,
+        user_id="memory_user",
+    )
+
+    # Mock the main agent
+    mock_agent = AsyncMock()
+    mock_agent.get_memory_info.return_value = {"blocks": 3, "tokens": 1500}
+    manager.main_agent = mock_agent
+    manager.active_agents["main"] = mock_agent
+
+    # Mock a sub-agent
+    mock_subagent = AsyncMock()
+    mock_subagent.get_memory_info.return_value = {"blocks": 2, "tokens": 800}
+    manager.active_agents["subagent_0"] = mock_subagent
+
+    # Get aggregated memory
+    memory_info = await manager.get_memory()
+
+    # Verify aggregation
+    assert len(memory_info) == 2
+    assert memory_info[0]["agent"] == "main"
+    assert memory_info[0]["memory"]["blocks"] == 3
+    assert memory_info[1]["agent"] == "subagent_0"
+    assert memory_info[1]["memory"]["blocks"] == 2
+
+
+@pytest.mark.asyncio
+@patch("services.chat_service.chat_agent.history_manager")
+async def test_end_to_end_orchestration(mock_history, orchestration_tools):
+    """Test end-to-end orchestration flow."""
+    mock_history.get_thread_history.return_value = []
+
+    manager = ChatAgentManager(
+        thread_id=106,
+        user_id="e2e_user",
+        tools=orchestration_tools,
+    )
+
+    # Mock the chat method on the main agent that will be created
+    with patch("services.chat_service.chat_agent.ChatAgent.chat") as mock_chat:
+        mock_chat.return_value = "Orchestrated response"
+
+        # This should build the system and process the query
+        response = await manager.chat("Hello, orchestrate this!")
+
+        assert response == "Orchestrated response"
+        assert manager.main_agent is not None
+        mock_chat.assert_called_once_with("Hello, orchestrate this!")
